Update bar component
